type: feat
Added error-counting capabilities to the application. Number
of incorrect chracters entered are now counted and added to
the stats screen which is displayed after the article has
been typed.

For now this error value is displayed and forgotten about-
eventually this can be incorporated along side WPM into
a database which will be queried to display more useful
statistics and trends regarding the users progress.